Task 2 ( Responsive  webpage)
As it is the second task of web development, created a responsive webpage with additional features in it. As it is created with the help of Html, CSS and little java script. As you can see here :      file:///C:/Users/K%20Gowthami/OneDrive/Desktop/mainflow%20intern/Task%202(responsive%20web%20page)/index.html